feat: breake js into modules
