+++ index.html
@@ -254,7 +254,6 @@
         </div>
       </article>
     </template>
-    <script src="./scripts/constants.js"></script>
-    <script src="./scripts/script.js"></script>
+    <script type="module" src="./scripts/script.js"></script>
   </body>
 </html>
